# docs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: AI Resume Parser API
  version: 1.0.0
  description: >
    API for uploading resumes, extracting structured data using AI models,
    storing parsed output, and matching resumes against job descriptions.

servers:
  - url: http://localhost:8000
    description: Local Dev Server

tags:
  - name: Resume
    description: Upload & parse resumes
  - name: Matching
    description: Resume vs Job description scoring
  - name: System
    description: Health check

paths:
  /api/v1/health:
    get:
      tags: [System]
      summary: Health Check
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string

  /api/v1/resumes/upload:
    post:
      tags: [Resume]
      summary: Upload and parse resume
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Resume file to upload
      responses:
        '200':
          description: Resume uploaded & parsed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /api/v1/match/{resume_id}:
    post:
      tags: [Matching]
      summary: Match parsed resume with job description
      security:
        - BearerAuth: []
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequest'
      responses:
        '200':
          description: Match score calculated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
        '404':
          description: Resume not found
        '500':
          description: Server error

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ParseResult:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            type: string

    UploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string
        data:
          type: object
          properties:
            resume_id:
              type: string
            extracted_data:
              $ref: '#/components/schemas/ParseResult'

    MatchRequest:
      type: object
      properties:
        job_description:
          type: string
      required:
        - job_description

    MatchResponse:
      type: object
      properties:
        resume_id:
          type: string
        score:
          type: number
          format: float
        keywords_matched:
          type: array
          items:
            type: string